frontend: changing-report-dialog.component.ts - functional for choosing timeslot and report title from drop list and filling default values from previously created report were added other files - tuning of styles and layout

+++ frontend/src/main/angular/src/app/componets/report-card/report-card.component.ts
@@ -88,9 +88,9 @@ export class ReportCardComponent {
         dialogRef.afterClosed().subscribe(result => {
             console.log("res of changing dialog")
             console.log(result);
-            if (result != null ) {
+            if (result) {
                 this.reportService.updateReport(result).subscribe(() => {
-                    this.reportDashboard.ngOnInit();// Перенаправление на текущую страницу
+                    this.reportDashboard.ngOnInit();
                 });
             }
         });
@@ -116,9 +116,12 @@ export class ReportCardComponent {
         dialogRef.afterClosed().subscribe(result => {
             console.log("res of canceling dialog")
             console.log(result);
-            this.reportService.deleteReport(result).subscribe(() => {
-                this.reportDashboard.ngOnInit();// Перенаправление на текущую страницу
-            });
+            if (result) {
+                this.reportService.deleteReport(result).subscribe(() => {
+                    this.reportDashboard.ngOnInit();
+
+                });
+            }
         });
     }
 }
